Move inner_timeout_ms def to kafka.util timeout_ms_fn
